requirements.txt
django
redis
requests
aiocoap
dnspython==1.16.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pyserial==3.4
sleekxmpp==1.3.2
